Feature: rating update

+++ src/main/java/excopen/backend/dto/ReviewResponseDTO.java
@@ -1,6 +1,8 @@
 package excopen.backend.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -8,7 +10,7 @@ public class ReviewResponseDTO {
     private Long id;
     private Long tourId;
     private Long userId;
-    private Integer rating;
+    private BigDecimal rating;
     private String reviewText;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
